add factories for power up blocks
