+++ TheAlienThatEatsTheCarrot/GameEngine/Entities/Factories/BlockFactories/PowerupBlockFactories/InvinciblePowerupBlockFactory.swift
@@ -1,0 +1,21 @@
+//
+//  InvinciblePowerupBlockFactory.swift
+//  TheAlienThatEatsTheCarrot
+//
+//  Created by Justin Cheah Yun Fei on 11/4/24.
+//
+
+import Foundation
+
+class InvinciblePowerupBlockFactory: PowerupBlockFactory {
+    override func createComponents() -> [Component] {
+        var powerupBlockComponents = super.createComponents()
+        let collisionEffect = SpawnPowerupCollisionEffect(powerupType: .invinsible, spawnDirection: .up)
+        let collisionEffectComponent = CollisionEffectComponent(entity: entity,
+                                                                acceptableComponentsColliders: [PlayerComponent.self, EnemyComponent.self],
+                                                                acceptableDirectionsToCollideFrom: [.down],
+                                                                collisionEffect: collisionEffect)
+        powerupBlockComponents.append(collisionEffectComponent)
+        return powerupBlockComponents
+    }
+}
